refactor(agent-bff): assemble the server behind a single buildBff
`runCli` owned the middleware order, so an embedding host had no way to get
the same stack without a listener. Move the whole assembly into `buildBff`,
which returns the request handler; `runCli` is now parseConfig + buildBff +
listen.

`/health` and the version header move with it, so they belong to the handler
rather than to the listener. `BFFHttpServer` keeps its `middlewares`
constructor working for external consumers and gains an optional `callback`.
