+++ packages/agent-bff/src/cli-dispatch.ts
@@ -6,7 +6,8 @@ import path from 'path';
 
 import createConsoleLogger from './adapters/console-logger';
 import { AI_QUERY_ROUTE } from './ai/ai-routes-middleware';
-import runCli, { resolveOAuthConfig, resolveUnfoldSource } from './cli-core';
+import { resolveOAuthConfig, resolveUnfoldSource } from './build-bff';
+import runCli from './cli-core';
 import { parseConfig, parsePublicUrl } from './config/env-config';
 import { extractErrorMessage } from './errors';
 import { generateOpenApiDocument, serializeOpenApi } from './openapi/openapi-document';
