+++ packages/agent-bff/src/index.ts
@@ -2,6 +2,8 @@ export { default as BFFHttpServer } from './http/bff-http-server';
 export { parseConfig, REQUIRED_KEYS } from './config/env-config';
 export type { BFFConfig, PresenceMap, RequiredKey } from './config/env-config';
 export { default as runCli } from './cli-core';
+export { default as buildBff } from './build-bff';
+export type { Bff, BuildBffOptions, BffCallback } from './build-bff';
 export { ConfigurationError } from './errors';
 export { default as DEFAULT_BFF_PORT } from './defaults';
 export { default as createConsoleLogger } from './adapters/console-logger';
